Added a "guess" mode for the MIME.Message constructor.
For some impossible reason though, it still manages to
throw an error for a message on the Roxen list. More
specifically, the error I didn't manage to fix is this:

invalid parameter "CHARSETiso-8859-1" in Content-Type
 "TEXT/PLAIN; CHARSETiso-8859-1" (0)
/usr/local/pike/7.3.47/lib/modules/MIME.pmod:857:
  create("Content-Type: TEXT/PLAIN; CHARSETiso-8859-1\n
  Content-Transfer-Encoding: 8BIT\nContent-ID: <Pine.GSO
  .4.21.0009131219142.19895@jms.idonex.se>\n\n----------
   Forwarded message ----------\nFrom: Son\351 Dia"+[858
  ],0,0,0)

for a call to MIME.Message(body, 0, 0, 1); the "(0)"
seems to hint that the guess parameter is dropped or
reset somewhere, but I can't figure out how nor where.

Rev: src/modules/MIME/module.pmod.in:1.28