Fixed duplicated destroy of the mutex condvar that affected macos and windows.
The behavior for mutexes is also slightly changed so that destructing a
mutex no longer destructs the key that is locking it. This is to make the
old de-facto behavior consistent where threads waiting to lock the mutex
could do that even after it has been destructed; only the thread currently
holding the lock would lose it asynchronously. The alternate PICKY_MUTEX
behavior is consistent the other way around by both destroying any
outstanding key and causing all wait calls to break immediately with an
exception. That is however disabled since it turns out to not work well with
old code that inherit mutexes in objects that are destructed explicitly.

Rev: src/threads.c:1.235