Encode references to all ex-masters as if they are to the current master.
That since we cannot hope to eradicate all references to the old
master when replacing it, and in any case the decoder cannot hope to
find the same ex-master again.

If this creates a problem for someone using replace_master then it's
always possible to clear the is_pike_master variable in the old
master.

Rev: lib/master.pike.in:1.476