Fixed stack overrun bug in _Crypto.crypto, _Crypto.cbc, and _Crypto.invert.
This bug could cause out of C-stack by feeding large data strings to crypt()
in these classes. With the exception of _Crypto.invert.crypt_block, it's not
a clear-cut stack smash though, since the result string was allocated on the
stack, and there should be several check_c_stack calls via safe_apply before
anything gets written to it.