Added support for typed constants in the doc extractor.
The type can either be given explicitly between the "constant" keyword and
the identifier, or it can be determined from the constant expression if it's
specified and is a plain literal.

Note: This makes the parsing of constant declarations somewhat more strict,
which can cause errors in existing AutoDoc markup.

This change makes the constant declarations in the GSSAPI module correct.