Improved typed mode for DECIMAL fields.
They are now returned as integers if they are declared to have no decimals,
otherwise they are returned as Gmp.mpq objects.

Compatibility note: This (along with the new handling of BIT fields) is an
incompatible change in typed mode, but there's probably not much code that
uses it yet, and there has been no official release with the old behavior.
It has been around in git for 1½ years, though.