Instantiate the fallback codecs instead of using the master directly.
(encode|decode)_value now instantiates a local codec instance from
master()->Encoder or master()->Decoder, respectively. That to make it
possible to keep local state. Specifically, it fixes the problem that the
reverse lookup mappings in the encoder get populated way too early when the
master is used directly.

The codecs are only instantiated on-demand, though.