Made some function non-PMOD_EXPORT. Added pure and malloc attributes.
This generates somewhat better code for some functions.

As an example, reverse(40Mb string) now takes 0.098 instead of 0.114s,
or about 16% faster on my computer (gcc 4.4.3).

We really should go through all functions that are exported, and see
which can be un-exported, and perhaps some more can be made static.