GTK2: Add -e to MAKE_FLAGS
The GTK2 Makefile does a recursive "$(MAKE) OBJS=... all" to override
the list of objects.  However, all in turn also makes a recursive make.
Thus the OBJS= override is lost unless it is picked up from the environment.
That appears to be the default in GNU make, but other makes need -e for
this.