TEST: Describe `COVILHÃ´, a city in Portugal
This is a test of UTF-8 handling in CodeLibrarian.  The first line
of the commit message should contain the word COVILHA between matched
single quotes, with a tilde above the A.

The committed file also has the same character sequence in the filename
and the contents, all encoded as UTF-8.