pgsql: Fix reconnect option; more efficient decoding of selectresult; support text connection mode in queries
Text mode connection support is possible through a connection option and/or
through a query option.  Turning this on will allow multiple statements
per query, but will also force communication to and from the database for
a query to be converted to text.  For some queries, this is more efficient
than the binary method.