[benchmark] Defeat the strength reduce optimizer in the loop tests
Added a new test with the number of iterations as a constant

I did this in order to get real values for iterations/second, which makes it
much easier to compare the relative speed of the tests.

This means that results from the old tests are not really relevant.

Also, currently this code:

for( int i=0; i<1; i++ ) n++;

generates 'i'=-1;n++;i=1' which does mean that each iteration
is somewhat slower than it should be.