[compiler][amd64] Real mov16 and mov8 added.
Using the movzx instruction, this is for unsigned numbers.  Versions
using movsz is needed if signed numbers are to be used.

Inlined a few more opcodes. Fixed branch when (non) zero and branch
when local to correctly treat 0.0 as non-zero.

Fixed clearing of zero type in ADD_LOCAL_INT[_and_pop] and
ADD_[NEG_]INT.