Java (NT): Use SetDllDirectory() to find required dlls. Fixes [bug 6471].
Oracle's jvm.dll has dependencies on runtime libraries that it doesn't
install in the global dll path or in the same directory. This patch
adds the directory where they do install the required dlls to the dll
search path, with a fallback to using the current directory on older NT.

The main change in this patch is to fix some calling-convention bugs in
earlier attempts, and to use the Unicode APIs.