Filesystem.Monitor: Fix race-condition when monitoring directories.
There was an over-caching issue in the directory monitors for
the following case (everything happend during the same second):

  1) A new directory is created.

  2) The monitor detects that the directory was created, and checks
     its contents -- empty.

  3) Some files are added to the directory.

When the monitor returned to check the directory a bit later, it
saw that the mtime was unchanged (since everything above happened
during the same second), so it didn't bother to check the contents.

Now the directory contents are rescanned anyway as long as the
directory hasn't been marked stable.

Also removed some priority bumping in check() that was already
performed properly by status_change().

Fixes [bug 6577] and possibly also some of [bug 6534].