SSL.sslfile: Avoid rescheduling the ssl_read_callback on no read_callback.
If ssl_read_callback() was called in reschedule mode, and there was
data in the read_buffer, it would be rescheduled again even if there
was no read_callback configured. This could lead to call_out loops
taking 100% cpu, since no data would be read from the read_buffer.

This patch avoids rescheduling if read_callback in the above case.

Potential fix for [bug 6582].