precompile: Support INHERIT ::this_program.
Added support for the syntax

  INHERIT ::this_program;

which has the same semantics as the corresponding syntax in Pike.

This is yet another step in simplifying inherits of code written
in Pike from C-modules. The typical use-case is where there is a
base class written in Pike where one of the nested classes needs
to be overloaded with an optimized version written in C.