Add function String.levenshtein_distance()
That function calculates the Levenshtein distance between two strings.
The Levenshtein distance describes the minimum number of edit operations
insert, delete or substitue a character to get from one string to the other.

The algorithm can be used in approximate string matching to find matches for
a short string in many longer texts, when a small number of differences is
expected.