[backend] Avoid busy-looping to get the the delay for call_out 100% right
It really kills the CPU, especially when you have a lot of open FD:s.

The closest 1ms is good enough, really, since it's not really possible
to guarantee anything better than that anyway (the time it takes to
run (e)poll() is rather varied, and usually in at least in the 0.1ms
range).