The operation "empty_string += string" is now significantly faster.
This is rather common in code that works with buffers. Previously it
was, somewhat surpisingly, significantly faster to write code like:

  if( strlen( x ) )
     x += str;
  else
     x = str;

than it was to simply add the str to x regardless of whether or not x
already had contents.

Now the check is done in `+() instead. However, it still does not
handle things like "" + str + "", which might perhaps be useful, but
is significantly less common in real code.

str + "" was already handled.