Be significantly more strict when checking vary function types.
The type of objects and ranges of strings and integers are now checked.

Still left to do is function arguments/return values for function
arguments.

Also, bignum numbers and other objects emulating integers (Gmp.mpz as
an example) are not currently 100% correctly checked.

An example that works now:

| class Test1 { int b(); }
| class Test2 { int a(); }
|
| class Foo
| {
|     variant string foo( object q ){ return "Genericfoo"; }
|     variant string foo( Test1 x ) { return "Test 1"; }
|     variant string foo( Test2 x ) { return "Test 2"; }
|
|     variant string bar( int(0..0) s ) { return "0"; }
|     variant float bar( int(1..1) s ) { return 1.0; }
|     variant int bar( int(2..2) s ) { return 2; }
| }

However, the last function in the class that match any given arguments
is always used.

There is no concept of 'best' match.

So, if the 'foo( object q )' variant had been last that one would
always have been used.