[performance] Do not use block-alloc for pike_frame and catch_context
They are too important for code execution speed.

struct pike_frames are allocated in chunks but not free:d until the
program exists.  This is basically just like the normal stack, and for
all but the most extreme of recursive programs this is not really an
issue. And for those programs the only loss now is that we are not
returning the frame memory to the system, we are actually using less
memory at peak.

The catch_context structures (that are fairly large, anyway, 80 bytes
on my machine) are simply allocated using malloc, and up to 100 free
ones are kept in a list for quick use.