Introduce support for padded blobs. This is done to counter the performance degradation seen when incrementally growing blobs which causes crazy record fragmentation in MySQL's word_hit.MYD file.
The strategy is to grow blobs 25% when full, and otherwise write in place
and adjust counters to handle book-keeping of the padding. As a test case
I had a document containing ~1100 words (50% unique words) and indexed it
repeatedly to a database with 620.000 records; the old implementation
deteriorated from 270 ms -> 1100 ms per pass after 200 passes (although
OPTIMIZE TABLE would bring it back down again), while the new one is more
or less stable at 270-300 ms per pass throughout the whole test.

Padded blobs are not enabled by default unless higher layers of the
application indicates that it is ok. This is tested by resolving the
Search.Process.Compactor class since that code depdends on the blob format
and if an older compactor is present we cannot make the switch. (This would
preferably be solved by moving the compactor to this pmod.) If no compactor
is present the behavior (as well as the word_hit table definition) is
unchanged from earlier versions.

Note that if padding is activated in a multi-server setup with a shared
database all clients must be upgraded to the same implementation. The record
format is not backwards compatible since old blob readers will fail to skip
the padding, and old clients will write records with incorrect padding
counters.