Disabled api-version system (or, well, default to max api version)
If we want to be able to compile old versions of pike with new
precompile.pike, and new versions of pike with old precompile.pike we
have to ignore the api version.

Since none of the API changes really change how previously valid code
is interpreted the whole API system only adds incompatibility.

As an example, backend.cmod in pike 7.8 can not be precompiled if you
only have a pike 7.9 installed, which makes it impossible to build a
pike 7.8 from git if you have pike 7.9+ installed (specifically, the
object(Thread.Thread) will break, since newer precompile added actual
support for it, while the backend.cmod uses the old non-support, but
both version works just fine if the check is removed).

Adding the --api argument to the 7.8 makefiles is not really an
option, because that will not work for older releases, and will make
it impossible to precompile pike 7.8 .cmod-files with pikes older than
7.9 (which is sort of non-optimal).

Similar issues occured when compiling pike 7.9 with an
installed pike 7.8.

Simply removing the api check makes all issues just go away.