Add PS_NAME attribute in info mapping if possible. Needed to facilitate compatibility with code relying on behaviour of older versions of FreeType.
Background: modern FreeType implementations return the FONT_FAMILY and
FONT_SUBFAMILY SFNT attributes as family and style in the mapping
returned by the info() function; older versions (at least FreeType
2.1.9 and older) returned PS_NAME and a string derived from other data
for those fields.

The problem was noticed when a Roxen customer wanted to update their
Roxen CMS installation from 4.5 (based on on Pike7.4 with
FreeType2.1.9) to a more modern version. Additional fixes in Roxen
will be required for full compatibility, but this Pike change makes
the necessary data available to enable that. (Roxen internal note:
RT#21337.)