Do not use alloca in sprintf
this is an attempt to fix a stack corruption bug which occurs in pike
versions compiled with msvc. the bug appears to be triggered by using
both setjmp and alloca in the same frame. even though sprintf does not
directly jump, its calling va_error, which does. in this patch alloca is
replaced by a simple temporary allocator.