Debug.Profiling and Debug.Watchdog
Debug.Profiling can be used to collect and format profiling
information.

It depends on pike being compiled with profiling support.

Debug.Watchdog is a class that can be instanced that ensures the pike
backend is actually running.

If it is not, or if SIGQUIT is received (or print_debug is called) it
will report all thread backtraces and various other debug information
to stderr, and, if the output was triggered by the watchdog, kill the
process.

It is also possible to run the watchdog in a backend-less mode, in
this mode it is up to the application to regularly call the ping()
function in the watchdog.