Merged the "ph/bits" branch.
This allows one more syntax for int-ranges in types:

(Xbit): This reprensents the range (0..(1<<X)-1).

So, string(8bit) is identical to string(0..255) and string(16bit) is
string(0..65535).

The same is true for integers, int(3bit) is int(0..7).

This does not conflict with any existing code because identifiers can
not start with a number.