Compat: 7.8::SSL: Improved robustness in destroy().
Applies the "SSL: Improved robustness in destroy()." fix from
Pike 7.8 in the Pike 7.8 compat SSL module:

The embedded stream may still be registered with a backend when the
sslfile object loses its last reference. Make sure not to trigger
the "In callback mode in a different backend." error in that case.

Fixes [bug 6958].