Build: Fixed default target override in semidynamic mode.
The default Makefile rule wasn't executed properly in
semidynamic modules. This was due to the mismatch between
$MODULE_BUILD_TYPE ("static") and the contents of build_type
("semidynamic") causing the "static" rule to be executed
instead of the default rule.

A typical example of an affected module is "Shuffler", where the
effect was that the generated module.a only contained the main
object file and not any of the other object files.