Compiler: Fix scope name for inherit ::this_program.
The scope name for inherits using ::this_program used to
be "this_program", which caused the "this_program::"-scope
to refer to the inherit rather than to the current class.

The scope now instead gets its name from the name of the current class.

Also fixes Crypto.CTR.State, which triggered the problem.