Standards.ASN1.Decode: Support reencoding.
Primitive and Constructed now have get_der() that returns the actual
der encoding of the object (via der_encode()).

Also adds get_contents() that returns the old value of get_der()
(ie just the payload), and a few getters to improve API compat
with the classes in Standards.ASN1.Types.