Silenced some compiler warnings.
One minor actual but very unimportant fix:

The backend wakeup callback could fail to read the data
from the wkeup pipe if it received a very well (for values
of well) timed signal.

However, the only effect would be that the backend is
immediately woken up again the next time it is started,
and clear the pipe buffer then instead.