Backend: Improved support for OOB with kqueue(2).
On MacOS X out-of-band data on sockets is signalled by the flag EV_OOBAND
(aka EV_FLAG1) in the EVFILT_READ notification. Unfortunately this
notification is by default only sent when there is also normal data
available. The kernel source indicates that it should be possible to get
notifications on just OOB by setting EV_OOBAND in the call to kevent(2)
(this is what poll(2) does internally). kevent(2) however masks the flag
before calling the internal function kevent_register(), so it is not
possible at this time.

On FreeBSD it seems out-of-band data is signalled as a normal EVFILT_READ
notification.

Improves the status for [bug 7063], but requires kernel changes
to fix the problem on MacOS X fully.