Optimized Standards.URI `== and _sprintf rather significantly.
Note that _sprintf is also used for cast-to-string (and was previously
used for `==).

For the most common case (Standards.URI(uri), where uri is a complete
url, which is then casted to a string or used as a mapping index) the
code is now about 10x faster.

Also added __hash.