SSL.context: Never sort null ciphers before non-null.
Cipher suites with server authentication, but no encryption were
preferred to suites with encryption, but no authentication. Now
the level of authentication is set to none if there's no encryption,
which means that the order now is reversed in this case.

Note that this only matters if you have called get_suites() with both
a first argument of -1 (include null ciphers) and a second argument
of 2 (include anonymous key exchange methods).