Compat: 7.8::SSL.sslfile: Make sure not to hang on destruct.
Applies the "SSL.sslfile: Make sure not to hang on destruct." fix from
Pike 7.8 in the Pike 7.8 compat SSL module:

The destruct callbac could hang in the local backend when attempting
to close the connection cleanly. Make sure that the local backend
doesn't wait for I/O completion in this case.

Fixes [bug 7066].