SSL.Cipher: Reduced DH overhead by a factor ~8.
The DH secret was generated ~8 times larger than needed and
then reduced to the required size. This provided for an
excellent distribution at a cost of quite a bit of overhead.

As this wasn't the intended behaviour of the code, it has
now been changed to just generate the required number of bits
plus a fixed overhead of 128 bits, which ought to be enough to
reduce the bias to a minimum.