Backend: Invalidate the call_out id when stale.
Reduce garbage by not keeping references to arguments for dead call_outs.

It is not unusual for a call_out user to keep the call_out id until
the next time the call_out is needed. As these quite often cause
circular references, we need to zap them when they're no longer needed.