calc(*) are now PCHARP functions.
I also fixed five errors that hopefully did not really matter all that
much in calc(*):

o Two instances of data[1] instead of data[pos+1]
o Three instances of data[pos]+1 instead of data[pos+1]

The first error would be triggered if you used '/' in an #if and the
second character of the current cpp work area was '/'.

The second one is probably harmless. Or not, it was in the code for
'&' and '|': basically the detection of && and || there was broken.

However, I think that would be handled by the calling order of the
calc* functions.

Anyway, cpp.o is now less than 100k, about 45% of its original size
(OSX 64-bit).