low*_cpp now uses PCHARP.
Shaves of about 50% of the remaining code size.  Now down to about
50Kb, which seems more reasonable than 200k.

There is still a lot of code duplication, though, like the places that
insert the current file as a #line, the code that handles quoted
newlines etc.

But it is better.

I also removed macros that were only used once and inlined them.