Nettle: Improve propagation of errors.
Use apply() instead of safe_apply() in lots of places. In addition
to improving the propagation of errors, it should reduce the stack
usage and improve performance, since it avoids the longjump contexts.

Also adds quite a few missing proxy functions, and improves the types
for some of the existing as well.