Added a few more limits (from man getrlimit on linux)
Also rewrote the limit handling to have the limits in an array.
This is somewhat slower, actually, since it is at least currently
using strcmp, but it is probbaly not much of an issue.

This rewrite also allows code reusage from signal_handler.c, so
that we support the same set of limits in both functions.