Fix for Opera bug TURO2-552
"Character 0x006f5242 at index 154 is outside the allowed range."

The source of the problem is entities along the lines of &#1928477;.

The bug is fixed by modifying the HTML entity decoder to behave like
the browsers do: All characters in a URL which are greater than the
limit for UTF8 characters i.e. > 0x10ffff will be replaced with the
character 0x0000fffd.