Removed the support for mmx.h
This was a header file that existed back in ye good old time, allowing
usage of MMX without much in the way of compiler support.

However, it is not really available anywhere any more, and almost all
code that used it was disabled.

If we do feel the need for MMX (or rather SSE3+, really, MMX is rather
old-school these days) we should use compiler intrinsics.