Image.PSD: Removed erroneous range check.
This check was the wrong way around for the common case where
the bit packer actually has packed some bits, but it can also
not be left in place as an inverted test, as the packed bits
may contain noop instructions, or more likely unpacked bits
(which will expand by a factor 129/128).

Note also that packbitsdecode() does contain range checks.

Fixes Image.decode_layers() on psd-files.