Backend: Fixed EXIT in CallOut.
EXIT in CallOut sometimes attempts to get the parent object (ie the backend)
by calling parent_storage(), but that requires destruct() to have registered
the program with a destroy_called_mark, which it only does for programs
lacking lfun::destroy() if they are

  * C-programs (no problem here).

and

  * Have the PROGRAM_NEEDS_PARENT flag set (this was missing).

This caused the error "Cannot access parent of destructed object.".

This bug could sometimes be triggered by the async_tls_close_test.